(acm) Currently Listening To: 8863220 SUDE4476 NULL YOUTUBE SUMMER DEALS FY27 XCAT SUDE EGBG VID 1920x1080 15 GM EL 26010930001PWW - Walmart

+++ src/libs/types/scoreboard.rs
@@ -6,7 +6,7 @@ use crate::libs::types::teams::TeamId;
 pub struct ScoreboardEntry {
     pub team_id: TeamId,
     pub team_name: String,
-    pub points: f64,
+    pub points: u32,
     pub last_solve_time: Option<i64>,
     pub solves: Vec<String>, // list of challenge ids
     pub rank: u32,
